requirements.txt
albumentations==1.1.0
albumentations==1.3.1
lightning==2.0.6
lightning==2.1.0rc0
matplotlib==3.5.3
matplotlib==3.4.1
matplotlib==3.2.2
matplotlib==3.0.0
matplotlib==3.2.1
matplotlib==3.5.1
matplotlib==3.6.2
matplotlib==3.4.3
matplotlib==3.3.4
matplotlib==3.7.2
numpy==1.21.6
numpy==1.25.2
numpy==1.23.4
numpy==1.18.1
numpy==1.22.4
numpy==1.21.4
numpy==1.21.0
numpy==1.25.0
numpy==1.24.3
numpy==1.19.2
numpy==1.20.1
numpy==1.17.0
numpy==1.21.1
numpy==1.22.3
numpy==1.20.3
numpy==1.19.5
numpy==1.18.5
numpy==1.24.4
Pillow==6.2.0
Pillow==6.2.1
Pillow==10.0.0
Pillow==8.4.0
Pillow==7.0.0
Pillow==9.3.0
Pillow==9.0.1
Pillow==8.3.1
Pillow==8.2.0
Pillow==8.1.2
Pillow==9.5.0
Pillow==9.4.0
Pillow==10.0.1
PyYAML==5.3.1
PyYAML==5.3
PyYAML==5.4.1
PyYAML==6.0.1
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
torch==2.0.0
torch==1.5.0
torch==1.5.1
torch==2.0.1
torch==1.8.0
torch==2.1.0.dev20230728
torch==1.11.0
torch==0.4.1.post2
torch==1.3.1
torch==1.0.1.post2
torch==1.4.0
torch==1.2.0
torch==1.9.0
torch==1.8.1
torch==2.0.1+cu118
torch==2.2.0.dev20231004
torchvision==0.15.1
torchvision==0.6.0
torchvision==0.15.2
torchvision==0.9.0
torchvision==0.4.2
torchvision==0.10.0
torchvision==0.16.0.dev20230728
torchvision==0.4.0a0+6b959ee
torchvision==0.2.2
torchvision==0.12.0
torchvision==0.5.0
torchvision==0.2.1
torchvision==0.9.1
torchvision==0.17.0.dev20231005
torchvision==0.15.2+cu118
tqdm==4.64.0
tqdm==4.47.0
tqdm==4.42.1
tqdm==4.62.2
tqdm==4.64.1
tqdm==4.62.0
tqdm==4.65.0
tqdm==4.46.1
tqdm==4.66.1
